# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Management API",
        "description": "API documentation for Hospital, Doctor, Patient, and Appointment management system",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/doctors": {
            "get": {
                "tags": [
                    "doctor"
                ],
                "summary": "Get all doctors",
                "operationId": "1ff2acc28425ccb6474cd7925c28ddd9",
                "responses": {
                    "200": {
                        "description": "List of doctors"
                    }
                }
            },
            "post": {
                "tags": [
                    "doctor"
                ],
                "summary": "Create a new doctor",
                "operationId": "0cfe89b6136ea425fda895fce9de82d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "specialization",
                                    "hospital_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "integer"
                                    },
                                    "specialization": {
                                        "type": "string",
                                        "format": "enum"
                                    },
                                    "Hospital_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Doctor created successfully"
                    }
                }
            }
        },
        "/api/v1/doctors/{id}": {
            "get": {
                "tags": [
                    "doctor"
                ],
                "summary": "Get a doctor by ID",
                "operationId": "873497a3e03b558bb6520c95313e6199",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the doctor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "doctor data retrieved"
                    },
                    "404": {
                        "description": "doctor not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "doctor"
                ],
                "summary": "Update a doctor",
                "operationId": "9d4a925ffd4e3b1007aaee415428f036",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the doctor to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "integer"
                                    },
                                    "specialization": {
                                        "type": "string",
                                        "format": "enum"
                                    },
                                    "Hospital_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "doctor updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "doctor"
                ],
                "summary": "Delete a doctor",
                "operationId": "d332a8941af19f68f3e7f32ee271e29e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the doctor to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "doctor deleted successfully"
                    }
                }
            }
        },
        "/api/v1/hospitals": {
            "get": {
                "tags": [
                    "Hospital"
                ],
                "summary": "Get all Hospitals Details",
                "operationId": "041c32c866f6521a62e8d306a4980751",
                "responses": {
                    "200": {
                        "description": "List of Hospitals"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hospital"
                ],
                "summary": "Create a new Hospital",
                "operationId": "2e53ae79f3d5f8006da137bc174c751d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "contact",
                                    "type",
                                    "services"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "contact": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "format": "enum"
                                    },
                                    "services": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hospital created successfully"
                    }
                }
            }
        },
        "/api/v1/hospitals/{id}": {
            "get": {
                "tags": [
                    "Hospital"
                ],
                "summary": "Get a Hospital by ID",
                "operationId": "978f234bf487cdc9c878ce520ce5917c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Hospital",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hospital data retrieved"
                    },
                    "404": {
                        "description": "Hospital not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hospital"
                ],
                "summary": "Update a Hospital",
                "operationId": "471d2ffdb43decb5c8e3fd5e283cdb85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Hospital to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "contact": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "format": "enum"
                                    },
                                    "services": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hospital updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hospital"
                ],
                "summary": "Delete a Hospital",
                "operationId": "1813cf095b22d4aaab318814a11f973a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Hospital to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Hospital deleted successfully"
                    }
                }
            }
        },
...
    },
...
}